Added lexical selection to Matxin.
This commit gives Matxin the ability to do lexical selection based on rules.
For now, there is a sample rule for es-eu at data/dict/eseu_lex.txt.

Also there has been a little bit of renaming on the prepositions selection code
in order to avoid conflicts and improve readability.